+++ app/src/main/java/com/createx/scalar/LogOut.java
@@ -1,15 +1,27 @@
 package com.createx.scalar;
 
 import android.content.Intent;
+import android.view.View;
+import android.widget.Button;
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 
 public class LogOut extends AppCompatActivity {
+    private Button logOut;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_log_out);
+
+        //from logout page to login
+        logOut = findViewById(R.id.back_to_start);
+        logOut.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                nextScreen(Login.class);
+            }
+        });
     }
 
     /**
